% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_earnings}
\alias{df_earnings}
\title{Example Median UK Salary and Deductions}
\format{
A data frame with 19 rows and 19 columns:
\describe{
  \item{date}{Covered period.}
  \item{annual_gross}{Annual gross income (in this example it's the median UK salary).}
  \item{monthly_gross}{Monthly gross income.}
  \item{monthly_net}{Monthly net (after subtracting all deductions) income.}
  \item{pensionable_earnings}{Earnings to calculate the pension deduction.}
  \item{pension_contribution}{Pension contribution.}
  \item{pension_deduction}{Pension deduction.}
  \item{national_insurance_allowance}{Monthly NI allowance (nothing is deducted below this value).}
  \item{national_insurance_contribution}{NI contribution, see \url{https://www.gov.uk/government/publications/rates-and-allowances-national-insurance-contributions/rates-and-allowances-national-insurance-contributions}.}
  \item{national_insurance_deduction}{NI deduction.}
  \item{tax_allowance}{Monthly tax allowance (no tax is paid on income below this value).}
  \item{tax_contribution}{Tax contribution.}
  \item{tax_deduction}{Tax deduction (it's calculated from the post-pension deduction income).}
  \item{student_loan_plan_2_allowance}{SLP2 allowance (a.k.a. plan 2 repayment threshold, see \url{https://www.gov.uk/repaying-your-student-loan/what-you-pay}).}
  \item{student_loan_plan_2_contribution}{SLP2 contribution.}
  \item{student_loan_plan_2_deduction}{SLP2 deduction.}
  \item{student_loan_postgrad_allowance}{SLP3 allowance (a.k.a. plan 3 repayment threshold, see \url{https://www.gov.uk/repaying-your-student-loan/what-you-pay}).}
  \item{student_loan_postgrad_contribution}{SLP3 contribution.}
  \item{student_loan_postgrad_deduction}{SLP3 deduction.}
}
}
\usage{
df_earnings
}
\description{
An artificial data with annual salary set to the median UK earnings in that year.
Originally, and Excel file allowing monthly deduction control, populated
from the monthly payslips.
}
\keyword{datasets}
