% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_exercises}
\alias{df_exercises}
\title{Example Exercises Data}
\format{
A data frame with 155 rows and 10 columns:
\describe{
  \item{date}{Date an exercises took place.}
  \item{exercise}{Exercise name.}
  \item{type}{Exercise type.}
  \item{category}{Exercise category.}
  \item{weight}{Total weight.}
  \item{sets}{Number of sets.}
  \item{reps}{Number of repetitions in a single set.}
  \item{reps_min}{Minimum number or repetitions.}
  \item{reps_max}{Maximum number or repetitions to increase weight.}
  \item{rest}{Rest time between sets.}
}
}
\usage{
df_exercises
}
\description{
Artificial exercises tracker data, originally an Excel file. Simple way of
seeing progress over time. The alternative would be to use some apps and export
the data instead.
}
\keyword{datasets}
