% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_historical_rpi}
\alias{df_historical_rpi}
\title{ONS Historical RPI Data}
\format{
A data frame with 73 rows and 2 columns:
\describe{
  \item{date}{Beginning of the month as date.}
  \item{rpi}{RPI in that period (in percentages).}
}
}
\source{
\url{https://www.ons.gov.uk/economy/inflationandpriceindices/timeseries/czbh/mm23}
}
\usage{
df_historical_rpi
}
\description{
A subset of the ONS RPI data.
}
\keyword{datasets}
