% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_issue_logs}
\alias{df_issue_logs}
\title{Example Issue Logs}
\format{
A data frame with 100 rows and 4 columns:
\describe{
  \item{content}{Log title.}
  \item{date}{When did the log take place?}
  \item{issue}{Issue the log relates to.}
  \item{log}{Log text.}
}
}
\usage{
df_issue_logs
}
\description{
Progress logs on resolving issues relating to personal development.
}
\keyword{datasets}
