% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_issue_relationships}
\alias{df_issue_relationships}
\title{Example Issues Relationships Data}
\format{
A data frame with 25 rows and 5 columns:
\describe{
  \item{from}{ID issue A.}
  \item{from_name}{Look-up of the A ID from the \code{df_issues} data.}
  \item{to}{ID issue B.}
  \item{to_name}{Look-up of the B ID from the \code{df_issues} data.}
  \item{value}{How much improving A improves B?}
}
}
\usage{
df_issue_relationships
}
\description{
Originally an Excel data showing how much improving issue A could improve issue B.
It's subjective and aims to capture the interconnected nature between many problems.
}
\keyword{datasets}
