% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_loan_repays}
\alias{df_loan_repays}
\title{Artificial Student Loan Repayments Data}
\format{
A data frame with 66 rows and 11 columns:
\describe{
  \item{date}{Begining of the month date.}
  \item{interest_rate_ug}{Interest rate for undergraduate loan in that month.}
  \item{interest_rate_pg}{Interest rate for postgraduate loan in that month.}
  \item{outlay_ug}{Money sent to the university covering your undergarduate tuition fees.}
  \item{outlay_pg}{Money sent to the university covering your postgarduate tuition fees.}
  \item{interest_ug}{Interest accrued for undergraduate loan in that month.}
  \item{interest_pg}{Interest accrued for postgraduate loan in that month.}
  \item{repays_ug}{Total undergraduate repayments (PAYE + overseas + voluntary) in that month.}
  \item{repays_pg}{Total postgraduate repayments (PAYE + overseas + voluntary) in that month.}
  \item{balance_ug}{Total undergraduate loan balance at the end of that month.}
  \item{balance_pg}{Total postgraduate loan balance at the end of that month.}
}
}
\usage{
df_loan_repays
}
\description{
Each financial year SLC sends a letter detailing student loan repayments and
interest accumulated over time. You can find example letters in the source code:
\href{https://github.com/Admate8/appPersonal/tree/main/inst/extdata}{\code{extdata}}.
\href{https://github.com/Admate8/appPersonal/tree/main/data-raw/student_loan_letters.R}{\code{student_loan_letters.R}} contains a script
allowing you to scrape this data and clean it. \code{df_loan_repays} data is
the result of such preprocessing.
}
\keyword{datasets}
