% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_measurements}
\alias{df_measurements}
\title{Example Measurements Data}
\format{
A data frame with 20 rows and 13 columns:
\describe{
  \item{date}{Measurement date.}
  \item{neck}{Neck measurement (cm).}
  \item{shoulders}{Shoulder measurement (cm).}
  \item{chest}{Chest measurement (cm).}
  \item{bicep}{Bicep measurement (cm).}
  \item{waist}{Waist measurement (cm).}
  \item{hip}{Hip measurement (cm).}
  \item{thigh}{Thigh measurement (cm).}
  \item{weight}{Weight measurement (kg).}
  \item{bmi}{BMI measurement.}
  \item{bmr}{Basal metabolic rate (calories).}
  \item{waist_to_hip_ratio}{waist / hip.}
  \item{target_calories}{1.1 * bmr.}
}
}
\usage{
df_measurements
}
\description{
Artificial measurements data over time.
}
\keyword{datasets}
