% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_nutrition}
\alias{df_nutrition}
\title{Example Nutrition Data}
\format{
A data frame with 700 rows and 12 columns:
\describe{
  \item{type}{Meal category, e.g. breakfast or dinner.}
  \item{cals}{Meal calories.}
  \item{fat}{Meal fat (g, 1g =~ 9cal).}
  \item{sat}{Meal saturates (g).}
  \item{carbs}{Meal carbohydrates (g, 1g =~ 4cal).}
  \item{fibre}{Meal fibre (g).}
  \item{sugar}{Meal sugars (g).}
  \item{prot}{Meal proteins (g, 1g =~ 4cal).}
  \item{sod}{Mead sodium (mg).}
  \item{chol}{Meal cholesterol (g).}
  \item{potassium}{Meal potassium (mg).}
  \item{date}{Meal date.}
}
}
\usage{
df_nutrition
}
\description{
The \href{https://www.fatsecret.com}{FatSecret} app is brilliant for tracking
meals and nutrition. It allows exporting your data as either PDF or CSV.
\href{https://github.com/Admate8/appPersonal/tree/main/data-raw/nutrition.R}{\code{nutrition.R}}, in the app source code, contains a script
I run weekly to collect and clean my data, which I then combine with historical
data.
}
\keyword{datasets}
