% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_rating}
\alias{df_rating}
\title{Example Exercises Rating Data}
\format{
A data frame with 69 rows and 3 columns:
\describe{
  \item{date}{Date an activity took place.}
  \item{rating}{How did you feel after the activity?}
  \item{type}{Type of activity.}
}
}
\usage{
df_rating
}
\description{
I like to measure how I feel after exercising and this data stores this info.
}
\keyword{datasets}
