% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_transactions}
\alias{df_transactions}
\title{Example Monthly Transactions Data}
\format{
A data frame with 1,053 rows and 5 columns:
\describe{
  \item{date}{Transaction date.}
  \item{category}{Transaction category.}
  \item{subcategory}{Transaction subcategory.}
  \item{description}{Transaction description.}
  \item{value}{Transaction value.}
}
}
\usage{
df_transactions
}
\description{
An artificial data with various transactions from March 2023 to the end of March 2024.
These transactions are always outgoing. Incoming transactions such as monthly
salary or money returns are not included.
}
\keyword{datasets}
