% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_utils_global_functions.R
\name{get_hex_colour_shade}
\alias{get_hex_colour_shade}
\title{Shade Hex Colour}
\usage{
get_hex_colour_shade(hex_color, factor = 0.2)
}
\arguments{
\item{hex_color}{String: hex colour.}

\item{factor}{Shading factor: negative/positive values will darken/lighten the colour, respectively.}
}
\value{
Shaded hex colour.
}
\description{
Shade Hex Colour
}
