% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_utils_global_functions.R
\name{get_rgb_colour}
\alias{get_rgb_colour}
\title{Convert HEX Colour to RGB Colour for the Echart Plots}
\usage{
get_rgb_colour(hex_color, opacity = 0.2)
}
\arguments{
\item{hex_color}{Hex colour.}

\item{opacity}{Opacity value between 0 and 1.}
}
\value{
Rgba string for the echart plots.
}
\description{
Convert HEX Colour to RGB Colour for the Echart Plots
}
