% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_utils_global_functions.R
\name{round_any}
\alias{round_any}
\title{Round to the Nearest Number}
\usage{
round_any(x, accuracy = 0.5)
}
\arguments{
\item{x}{Numeric vector.}

\item{accuracy}{The number to round to.}
}
\value{
Rounded vector of values
}
\description{
Round to the Nearest Number
}
