% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_utils_global_functions.R
\name{with_tooltip}
\alias{with_tooltip}
\title{Add a Tooltip to a Table Header}
\usage{
with_tooltip(value, tooltip, ...)
}
\arguments{
\item{value}{Table header title.}

\item{tooltip}{Tooltip content.}

\item{...}{Not in use.}
}
\description{
Add a Tooltip to a Table Header
}
