% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_loan_repays}
\alias{df_loan_repays}
\title{Artificial Student Loan Repayments Data}
\format{
A data frame with 66 rows and 11 columns:
\describe{
  \item{date}{Begining of the month date.}
  \item{interest_rate_xx}{Interest rate for either undergraduate (_ug) or postgraduate (_pg) loan in that month.}
  \item{outlay_xx}{Money sent to the university covering your undergarduate (_ug) or postgraduate (_pg) tuition fees.}
  \item{interest_xx}{Interest accrued for either undergraduate (_ug) or postgraduate (_pg) loans in that month.}
  \item{repays_xx}{Total repayments (PAYE + overseas + voluntary) in that month.}
}
}
\usage{
df_loan_repays
}
\description{
Each financial year SLC sends a letter detailing student loan repayments and
interest accumulated over time. You can find example letters in the source code:
\href{https://github.com/Admate8/appPersonal/tree/main/inst/extdata}{\code{extdata}}.
\href{https://github.com/Admate8/appPersonal/tree/main/data-raw/student_loan_letters.R}{\code{student_loan_letters.R}} contains a script
allowing you to scrape this data and clean it. \code{df_loan_repays} data is
the result of such preprocessing.
}
\keyword{datasets}
